Add CRUD functionality for Learner and ClassBatch objects. This task only concerns the backend, we will utilize some of it in later tasks.
Modify the Learner model so that a Learner object can be associated with multiple ClassBatch objects. Maintain the existing learner-class associations in this transition.

Add a “secret_id” field to the Learner model and ensure that it will not be returned via the GET Learner api endpoint.

Modify the ClassBatch page so that it displays associated Learner objects.

Add a page for viewing All Learners. Add a button on the Welcome page linking to it. For each Learner, add a delete button that would delete the Learner.

Add a page for viewing All ClassBatches. Add a button on the front page linking to it.